{{ cookiecutter.project_slug }}/pyproject.toml
[tool.poetry]
name = "{{ cookiecutter.project_slug }}"
version = "0.1.0"
description = "{{ cookiecutter.project_short_description }}"
authors = ["{{ cookiecutter.github_username }} <{{ cookiecutter.email }}>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.2"
djangorestframework = "3.14.0"
drf-spectacular = {extras = ["sidecar"], version = "^0.26.3"}

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[[tool.poetry.source]]
name = "aliyun"
priority = "default"
url = "http://mirrors.aliyun.com/pypi/simple/"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
